pyproject.toml
[project]
name = "dig-doc"
version = "0.0.1"
description = "dig-doc"
authors = []
dependencies = [
    "langchain>=0.3.7",
    "langchain-community>=0.3.5",
    "langchain-chroma>=0.1.4",
    "langchain-openai>=0.2.6",
    "unstructured[md]>=0.11.8",
    "nltk==3.8.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
]
requires-python = "==3.13.*"
readme = "README.md"

[project.optional-dependencies]
test = [
    "pytest>=8.3.3",
]
lint = [
    "ruff>=0.7.3",
    "mypy>=1.13.0",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
qa = {cmd = "python scripts/qa.py", env = { PYTHONPATH = "." }}
recreate-vectorstore = {cmd = "python scripts/recreate_vectorstore.py", env = { PYTHONPATH = "." }}

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "EM101",
    "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["INP001"]
"tests/*.py" = [
    "S101",
    "S105",
]
